// src/app/api/model/personCombinedCredits.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PersonCombinedCrewCredit } from './personCombinedCrewCredit';
import { PersonCombinedCastCredit } from './personCombinedCastCredit';


export interface PersonCombinedCredits { 
    cast?: Array<PersonCombinedCastCredit>;
    crew?: Array<PersonCombinedCrewCredit>;
    id?: number;
}