// src/app/api/model/taggedImagePage.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaggedImage } from './taggedImage';


export interface TaggedImagePage { 
    id?: number;
    page?: number;
    results?: Array<TaggedImage>;
    total_pages?: number;
    total_results?: number;
}